1.7v why it stops at 7%??

+++ beakJoon/greedyAlgorithm/beakjoon1202/beakjoon1202_usingHeap.c
@@ -107,7 +107,7 @@ int main(){
 			j++;
 		}
 		if(heapSize != 0){
-			answer += getMaxFromHeap();
+			answer += (long long int)getMaxFromHeap();
 		}
 	}
